refactor Windows transport host and local modules
