+++ litebox_broker_transport_windows_userland/src/named_pipe.rs
@@ -3,36 +3,25 @@
 
 //! Broker setup channels over Windows named pipes.
 
-use std::ffi::OsStr;
-use std::fs::{File, OpenOptions};
-use std::io::{Error, ErrorKind, Result as IoResult};
-use std::os::windows::fs::OpenOptionsExt;
-use std::os::windows::io::{AsRawHandle, FromRawHandle};
-use std::time::{Duration, Instant};
+mod host;
+mod local;
 
-use litebox_broker_protocol::message::{BrokerHandshakeRequest, BrokerHandshakeResponse};
-use litebox_broker_protocol::wire::{
-    WireError, decode_handshake_request, decode_handshake_response, encode_handshake_request,
-    encode_handshake_response,
-};
-use litebox_broker_transport::channel::{
-    HostReceive, HostSetupChannel, LocalSetupChannel, PeerCredential,
-};
-use litebox_broker_transport::control_ring::ControlRing;
+pub use host::{WindowsNamedPipeHostSetupChannel, validate_client_process};
+pub use local::WindowsNamedPipeLocalSetupChannel;
+
+use std::ffi::OsStr;
+use std::fs::File;
+use std::io::{Error, ErrorKind, Result as IoResult};
+use std::os::windows::io::{AsRawHandle, FromRawHandle};
 use windows_sys::Win32::Foundation::{
     ERROR_IO_INCOMPLETE, ERROR_IO_PENDING, ERROR_PIPE_CONNECTED, HANDLE, INVALID_HANDLE_VALUE,
 };
 use windows_sys::Win32::Storage::FileSystem::{FILE_FLAG_OVERLAPPED, PIPE_ACCESS_DUPLEX};
 use windows_sys::Win32::System::Pipes::{
-    ConnectNamedPipe, CreateNamedPipeW, GetNamedPipeClientProcessId, PIPE_READMODE_BYTE,
-    PIPE_TYPE_BYTE, PIPE_WAIT,
+    ConnectNamedPipe, CreateNamedPipeW, PIPE_READMODE_BYTE, PIPE_TYPE_BYTE, PIPE_WAIT,
 };
 
-use crate::setup::{OverlappedOperation, invalid_data, read_frame, wire_error, write_frame};
-use crate::shared_memory::{TransferredSharedMemory, WindowsSharedMemory};
-
-const CONNECT_RETRY_DELAY: Duration = Duration::from_millis(10);
-const TRANSFER_FRAME_TAG: u8 = 0xa1;
+use crate::setup::OverlappedOperation;
 
 /// One server-side named-pipe instance waiting for a single broker peer.
 pub struct WindowsNamedPipeListener {
@@ -42,21 +31,6 @@ pub struct WindowsNamedPipeListener {
 
 /// One connected named-pipe stream opened for overlapped I/O.
 pub struct WindowsNamedPipeStream(File);
-
-/// Local-side broker setup channel over a Windows named pipe.
-pub struct WindowsNamedPipeLocalSetupChannel {
-    stream: WindowsNamedPipeStream,
-    setup_deadline: Option<Instant>,
-    negotiated: bool,
-}
-
-/// Host-side broker setup channel over a Windows named pipe.
-pub struct WindowsNamedPipeHostSetupChannel {
-    stream: WindowsNamedPipeStream,
-    peer_credential: PeerCredential,
-    setup_deadline: Option<Instant>,
-    negotiated: bool,
-}
 
 impl WindowsNamedPipeListener {
     /// Creates one duplex byte-mode named-pipe instance.
@@ -145,241 +119,6 @@ impl WindowsNamedPipeStream {
     }
 }
 
-impl WindowsNamedPipeLocalSetupChannel {
-    /// Connects to a broker setup pipe until `deadline` expires.
-    pub fn connect_with_setup_deadline(name: &OsStr, deadline: Instant) -> IoResult<Self> {
-        loop {
-            match OpenOptions::new()
-                .read(true)
-                .write(true)
-                .custom_flags(FILE_FLAG_OVERLAPPED)
-                .open(name)
-            {
-                Ok(stream) => {
-                    return Ok(Self {
-                        stream: WindowsNamedPipeStream(stream),
-                        setup_deadline: Some(deadline),
-                        negotiated: false,
-                    });
-                }
-                Err(error) if Instant::now() < deadline => {
-                    std::thread::sleep(
-                        CONNECT_RETRY_DELAY.min(deadline.saturating_duration_since(Instant::now())),
-                    );
-                    let _ = error;
-                }
-                Err(error) => return Err(error),
-            }
-        }
-    }
-
-    /// Receives and maps one shared-memory object duplicated into this process.
-    pub fn receive_shared_memory(
-        &mut self,
-        expected_length: usize,
-    ) -> IoResult<WindowsSharedMemory> {
-        let frame =
-            read_frame(file_handle(&self.stream), self.setup_deadline)?.ok_or_else(|| {
-                Error::new(
-                    ErrorKind::UnexpectedEof,
-                    "broker closed before transferring shared memory",
-                )
-            })?;
-        let transfer = decode_transfer(&frame)?;
-        if transfer.length != expected_length {
-            return Err(invalid_data("transferred shared-memory length mismatch"));
-        }
-        // SAFETY: The authenticated broker duplicated these handles into this process and encoded
-        // their target-process values in the setup frame.
-        unsafe { WindowsSharedMemory::from_transferred(transfer) }
-    }
-
-    /// Receives and maps the shared control ring duplicated into this process.
-    pub fn receive_control_ring(&mut self) -> IoResult<WindowsSharedMemory> {
-        let frame =
-            read_frame(file_handle(&self.stream), self.setup_deadline)?.ok_or_else(|| {
-                Error::new(
-                    ErrorKind::UnexpectedEof,
-                    "broker closed before transferring control-ring memory",
-                )
-            })?;
-        let transfer = decode_transfer(&frame)?;
-        // SAFETY: The authenticated broker duplicated these handles into this process and encoded
-        // their target-process values in the setup frame.
-        unsafe { WindowsSharedMemory::control_ring_from_transferred(transfer) }
-    }
-
-    /// Activates calls and notifications over the transferred shared control ring.
-    pub fn into_active(
-        self,
-        ring: ControlRing<WindowsSharedMemory>,
-    ) -> IoResult<(
-        crate::control_ring::WindowsControlRingLocalCallChannel,
-        crate::control_ring::WindowsControlRingLocalNotificationChannel,
-    )> {
-        crate::control_ring::activate_local(self.stream, self.negotiated, self.setup_deadline, ring)
-    }
-}
-
-impl WindowsNamedPipeHostSetupChannel {
-    /// Creates a host setup channel after the deployment has authenticated the client.
-    pub const fn from_host_guaranteed(
-        stream: WindowsNamedPipeStream,
-        setup_deadline: Instant,
-    ) -> Self {
-        Self {
-            stream,
-            peer_credential: PeerCredential::HostGuaranteed,
-            setup_deadline: Some(setup_deadline),
-            negotiated: false,
-        }
-    }
-
-    /// Duplicates one shared-memory object into the runner and sends its handle values.
-    pub fn send_shared_memory(
-        &mut self,
-        memory: &WindowsSharedMemory,
-        runner_process: HANDLE,
-    ) -> IoResult<()> {
-        let transfer = memory.duplicate_to_process(runner_process)?;
-        write_frame(
-            file_handle(&self.stream),
-            &encode_transfer(&transfer)?,
-            self.setup_deadline,
-        )
-    }
-
-    /// Activates host request, response, and notification control-ring endpoints.
-    pub fn into_active(
-        self,
-        ring: ControlRing<WindowsSharedMemory>,
-    ) -> IoResult<(
-        crate::control_ring::WindowsControlRingHostRequestSource,
-        crate::control_ring::WindowsControlRingHostResponseSink,
-        crate::control_ring::WindowsControlRingHostNotificationChannel,
-        crate::control_ring::WindowsControlRingHostShutdown,
-    )> {
-        crate::control_ring::activate_host(self.stream, self.negotiated, self.setup_deadline, ring)
-    }
-}
-
-impl LocalSetupChannel for WindowsNamedPipeLocalSetupChannel {
-    type Error = Error;
-
-    fn send_handshake_request(&mut self, request: &BrokerHandshakeRequest) -> IoResult<()> {
-        write_frame(
-            file_handle(&self.stream),
-            &encode_handshake_request(request.clone()),
-            self.setup_deadline,
-        )
-    }
-
-    fn recv_handshake_response(&mut self) -> IoResult<Option<BrokerHandshakeResponse>> {
-        let response = read_frame(file_handle(&self.stream), self.setup_deadline)?
-            .map(|frame| decode_handshake_response(&frame).map_err(wire_error))
-            .transpose()?;
-        self.negotiated = matches!(response, Some(BrokerHandshakeResponse::Negotiated { .. }));
-        Ok(response)
-    }
-}
-
-impl HostSetupChannel for WindowsNamedPipeHostSetupChannel {
-    type Error = Error;
-
-    fn peer_credential(&self) -> IoResult<PeerCredential> {
-        Ok(self.peer_credential)
-    }
-
-    fn recv_handshake_request(&mut self) -> IoResult<HostReceive<BrokerHandshakeRequest>> {
-        let Some(frame) = read_frame(file_handle(&self.stream), self.setup_deadline)? else {
-            return Ok(HostReceive::PeerClosed);
-        };
-        match decode_handshake_request(&frame) {
-            Ok(request) => Ok(HostReceive::Message(request)),
-            Err(WireError::WrongMessagePhase) => Ok(HostReceive::ProtocolViolation),
-            Err(error) => Err(wire_error(error)),
-        }
-    }
-
-    fn send_handshake_response(&mut self, response: &BrokerHandshakeResponse) -> IoResult<()> {
-        write_frame(
-            file_handle(&self.stream),
-            &encode_handshake_response(response.clone()),
-            self.setup_deadline,
-        )?;
-        self.negotiated = matches!(response, BrokerHandshakeResponse::Negotiated { .. });
-        Ok(())
-    }
-}
-
-/// Validates that a connected named-pipe client belongs to `expected_process_id`.
-pub fn validate_client_process(
-    stream: &WindowsNamedPipeStream,
-    expected_process_id: u32,
-) -> IoResult<()> {
-    let mut process_id = 0;
-    // SAFETY: `stream` is a connected named pipe and `process_id` is writable.
-    if unsafe { GetNamedPipeClientProcessId(file_handle(stream), &raw mut process_id) } == 0 {
-        return Err(Error::last_os_error());
-    }
-    if process_id != expected_process_id {
-        return Err(Error::new(
-            ErrorKind::PermissionDenied,
-            "named-pipe client is not the expected runner process",
-        ));
-    }
-    Ok(())
-}
-
-fn encode_transfer(transfer: &TransferredSharedMemory) -> IoResult<Vec<u8>> {
-    let mut frame = Vec::with_capacity(13 + transfer.handles.len() * 8);
-    frame.push(TRANSFER_FRAME_TAG);
-    frame.extend_from_slice(
-        &u64::try_from(transfer.length)
-            .map_err(|_| invalid_data("shared-memory length is too large"))?
-            .to_le_bytes(),
-    );
-    frame.extend_from_slice(
-        &u32::try_from(transfer.handles.len())
-            .map_err(|_| invalid_data("too many transferred handles"))?
-            .to_le_bytes(),
-    );
-    for handle in &transfer.handles {
-        frame.extend_from_slice(
-            &u64::try_from(*handle)
-                .map_err(|_| invalid_data("transferred handle is too large"))?
-                .to_le_bytes(),
-        );
-    }
-    Ok(frame)
-}
-
-fn decode_transfer(frame: &[u8]) -> IoResult<TransferredSharedMemory> {
-    if frame.len() < 13 || frame[0] != TRANSFER_FRAME_TAG {
-        return Err(invalid_data("invalid shared-memory transfer frame"));
-    }
-    let length = usize::try_from(u64::from_le_bytes(frame[1..9].try_into().unwrap()))
-        .map_err(|_| invalid_data("transferred shared-memory length is too large"))?;
-    let count = usize::try_from(u32::from_le_bytes(frame[9..13].try_into().unwrap()))
-        .expect("u32 handle count fits usize");
-    if frame.len()
-        != 13
-            + count
-                .checked_mul(8)
-                .ok_or_else(|| invalid_data("handle count overflow"))?
-    {
-        return Err(invalid_data("invalid shared-memory transfer frame length"));
-    }
-    let handles = frame[13..]
-        .chunks_exact(8)
-        .map(|bytes| {
-            usize::try_from(u64::from_le_bytes(bytes.try_into().unwrap()))
-                .map_err(|_| invalid_data("transferred handle is too large"))
-        })
-        .collect::<IoResult<Vec<_>>>()?;
-    Ok(TransferredSharedMemory { length, handles })
-}
-
 fn wide_string(value: &OsStr) -> IoResult<Vec<u16>> {
     use std::os::windows::ffi::OsStrExt as _;
     let mut value = value.encode_wide().collect::<Vec<_>>();
@@ -401,20 +140,25 @@ fn file_handle(stream: &WindowsNamedPipeStream) -> HANDLE {
 mod tests {
     use super::*;
     use std::sync::{Arc, Barrier};
+    use std::time::{Duration, Instant};
 
-    use crate::setup::{OwnedEvent, read_pipe_until_cancelled};
+    use crate::setup::{OwnedEvent, read_pipe_until_cancelled, write_frame};
+    use crate::shared_memory::WindowsSharedMemory;
     use litebox_broker_protocol::message::{
-        BrokerNotification, BrokerOperation, BrokerRequest, BrokerResponse, BrokerResult,
-        ReadinessNotification,
+        BrokerHandshakeRequest, BrokerHandshakeResponse, BrokerNotification, BrokerOperation,
+        BrokerRequest, BrokerResponse, BrokerResult, ReadinessNotification,
     };
     use litebox_broker_protocol::readiness::ReadinessFlags;
     use litebox_broker_protocol::{BROKER_PROTOCOL_VERSION, ObjectHandle, RequestId};
     use litebox_broker_transport::channel::{
-        HostNotificationChannel, LocalCallChannel, LocalNotificationChannel,
+        HostNotificationChannel, HostReceive, HostSetupChannel, LocalCallChannel,
+        LocalNotificationChannel, LocalSetupChannel,
     };
     use litebox_broker_transport::control_ring::ControlRing;
     use litebox_broker_transport::shared_memory::SharedMemory as _;
     use windows_sys::Win32::System::Threading::{GetCurrentProcess, GetCurrentProcessId};
+
+    const ACCEPT_RETRY_DELAY: Duration = Duration::from_millis(10);
 
     fn pipe_name(suffix: &str) -> std::ffi::OsString {
         format!(r"\\.\pipe\litebox-broker-test-{}-{suffix}", unsafe {
@@ -454,7 +198,7 @@ mod tests {
                 Err(error) if error.kind() == ErrorKind::WouldBlock => {}
                 Err(error) => panic!("failed to accept named-pipe client: {error}"),
             }
-            std::thread::sleep(CONNECT_RETRY_DELAY);
+            std::thread::sleep(ACCEPT_RETRY_DELAY);
         }
     }
